backend/requirements.txt
Django>=4.2
mysqlclient>=2.2
paho-mqtt>=1.6
djangorestframework>=3.14
python-dotenv>=1.0
django-filter>=23.0
